requirements.txt
pandas
python-docx
openpyxl
pyperclip